+++ quickstart/eabm_package/single_element_partition.py
@@ -1,1 +1,1 @@
-../../tutorials/eabm_package/abaqus/single_element_partition.py
+../../tutorials/eabm_package/argparse_types/single_element_partition.py