+++ tutorials/eabm_package/argparse_types.py
@@ -1,0 +1,19 @@
+def positive_float(argument):
+    """Type function for argparse - positive floats
+
+    Abaqus Python 2 and Python 3 compatible argparse type method:
+    https://docs.python.org/3.8/library/argparse.html#type.
+
+    :param str argument: string argument from argparse
+
+    :returns: argument
+    :rtype: float
+    """
+    MINIMUM_VALUE = 0.0
+    try:
+        argument = float(argument)
+    except ValueError:
+        raise argparse.ArgumentTypeError("invalid float value: '{}'".format(argument))
+    if not argument > MINIMUM_VALUE:
+        raise argparse.ArgumentTypeError("invalid positive float: '{}'".format(argument))
+    return argument