+++ docs/changelog.rst
@@ -12,11 +12,18 @@ New Features
 ============
 - Add configuration files to the archive tutorial and quickstart archival task (:issue:`369`, :merge:`423`). By `Kyle
   Brindley`_.
+- Add positive-float input verification to the tutorial and quickstart files (:issue:`375`, :merge:`424`). By `Kyle
+  Brindley`_.
 
 Bug fixes
 =========
 - Fix issue in abaqus_file_parser where first frame of field output didn't get all the dimensions
   (:issue:`376`, :merge:`425`). By `Prabhu Khalsa`_.
+
+Documentation
+=============
+- Add a supplemental tutorial for input verification using `Argparse type`_ user-defined methods (:issue:`375`,
+  :merge:`424`). By `Kyle Brindley`_.
 
 ******************
 0.5.9 (2022-12-14)